feat(linter): wire preprocess.Scan in and migrate no-bare-urls (#337 Phase 2)
collectErrors now runs preprocess.Scan once per file and passes the
shared *preprocess.Context to rules that consume it. To avoid the memory
cost of reconstructing the lines slice per rule, un-migrated rules keep
receiving the raw lines (which Scan borrows, no copy); only no-bare-urls,
the reference adoption, takes the Context.

no-bare-urls drops its bespoke fence/comment tracking: it skips lines the
scanner classified as fenced code, indented code, HTML block, or HTML
comment, and scans the inline-sanitized text for the rest. This also
closes the indented-code and HTML-block gaps the previous implementation
missed (audit #337).

The inline-sanitization helpers (stripInlineCode, stripHTMLComments,
countBacktickRun) move to inline_strip.go: no-bare-urls no longer uses
them, but no-empty-links, link-fragments, consistent-emphasis-style,
no-hard-tabs, and fenced-code-language still do until Phase 3.

Refs #337 #340
